.github/workflows: Install WeasyPrint from `pypi.org`
This commit makes changes which install WeasyPrint via the Python
Package Index (PyPI) instead of via `apk`, the package manager of
Alpine Linux.

The reason for this is that the version of WeasyPrint installed via
`apk` does not match the current release of WeasyPrint in the official
WeasyPrint GitHub repository [1]. This makes it difficult to version
the provided Docker images in a consistent way.

To fix the problem, WeasyPrint is now installed via PyPI. PyPI provides
a simple interface [2] to query the latest version of WeasyPrint.

Nevertheless, using PyPI also requires more storage space for the
resulting Docker image. Previously, this could be avoided by using
Alpine Linux. But getting WeasyPrint from PyPI while using the
`python:alpine` Docker image is not possible because Alpine Linux uses
`musl libc` instead of `glibc and friends` [3]. This leads to problems
when installing Python packages, including WeasyPrint.

Link: https://github.com/Kozea/WeasyPrint [1]
Link: https://pypi.org/pypi/weasyprint/json [2]
Link: https://github.com/docker-library/docs/tree/master/python#pythonversion-alpine [3]
Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>