Move repository to `custom-images/weasyprint`
This commit moves the GitHub repository from 'weasyprint/weasyprint'
to 'custom-images/weasyprint'. The reason for this change is twofold:

Firstly, the name 'weasyprint' implies an official connection to the
WeasyPrint project. However, this is a privately managed Docker image
and there is no such connection.

Secondly, using the namespace 'custom-images' allows for the possibility
of providing additional Docker images through 'ghcr.io' in the future.

Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>